backend-ml/models/quality_model.py
from typing import Dict
from io import BytesIO
from PIL import Image, ImageStat

class ProduceQualityScorer:
    """Lightweight placeholder scorer. Replace with YOLOv5 integration for production accuracy."""

    def __init__(self):
        self.model_version = 'v0.1-placeholder'

    def _analyze_color(self, img: Image.Image) -> Dict:
        stat = ImageStat.Stat(img)
        mean = stat.mean[:3]
        brightness = sum(mean) / (255 * 3) * 10
        saturation = max(mean) / 255 * 10
        return { 'brightness': round(brightness,2), 'saturation': round(saturation,2) }

    def score_image(self, path: str) -> Dict:
        img = Image.open(path).convert('RGB')
        return self._score(img)

    def score_image_bytes(self, data: bytes, filename: str = None) -> Dict:
        img = Image.open(BytesIO(data)).convert('RGB')
        return self._score(img)

    def _score(self, img: Image.Image) -> Dict:
        color = self._analyze_color(img)
        # simple heuristic: higher brightness & saturation => higher quality
        base = (color['brightness'] + color['saturation']) / 20 * 100
        quality_score = max(0.0, min(100.0, base))
        return {
            'quality_score': round(quality_score,2),
            'confidence': 0.55,
            'defects': [],
            'color_analysis': color,
            'model_version': self.model_version
        }
