backend-ml/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
supabase==2.0.0
pillow==10.0.0
numpy==1.24.3
